move reasons and grade generating to the templates instead
