+++ shared/js/ui/templates/grade-scorecard.es6.js
@@ -6,9 +6,9 @@ const ratingHero = require('./shared/rating-hero.es6.js')
 module.exports = function () {
   return bel`<section class="sliding-subview sliding-subview--has-fixed-header">
     <div class="site-info site-info--full-height card">
-      ${ratingHero(this.model.site, { showClose: true })}
-      ${reasons(this.model.getReasons())}
-      ${grades(this.model.getGrades())}
+      ${ratingHero(this.model, { showClose: true })}
+      ${reasons(this.model)}
+      ${grades(this.model)}
     </div>
   </section>`
 }
