+++ shared/js/ui/views/site.es6.js
@@ -1,5 +1,4 @@
 const Parent = window.DDG.base.View
-const GradeScorecardModel = require('./../models/grade-scorecard.es6.js')
 const GradeScorecardView = require('./../views/grade-scorecard.es6.js')
 const GradeDetailsView = require('./../views/grade-details.es6.js')
 const PrivacyPracticesView = require('./../views/privacy-practices.es6.js')
@@ -95,13 +94,9 @@ Site.prototype = window.$.extend({},
     },
 
     _showGradeScorecard: function () {
-      const model = new GradeScorecardModel({
-        site: this.model
-      })
-
       this.views.gradeScorecard = new GradeScorecardView({
         template: gradeScorecardTemplate,
-        model
+        model: this.model
       })
     }
   }
